% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string.R
\name{char.get_between_separators}
\alias{char.get_between_separators}
\title{Return everything between first occurence of first separator and first occurence
of second separator after the first separator.}
\usage{
char.get_between_separators(string = "", first_separator = ",",
  second_separator = ",", include_sep_in_output = FALSE)
}
\arguments{
\item{string}{A character string.}

\item{first_separator}{Optional. Default ",".}

\item{second_separator}{Optional. Default ",".}

\item{include_sep_in_output}{Optional. Include the separators in the output. Default FALSE.}
}
\description{
Return everything between first occurence of first separator and first occurence
of second separator after the first separator.
}
\examples{
str <- "I am smart enough, good looking enough, and gosh darn it; people like me."
get_between_separators(str)
[1] "good looking enough"
char.get_between_separators(str, second_separator = ";")
[1] "good looking enough, and gosh darn it"
}
