% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string.R
\name{char.string2vector}
\alias{char.string2vector}
\title{Convert string to character vector}
\usage{
char.string2vector(string, delimeter = ",")
}
\arguments{
\item{string}{A character string.}

\item{delimeter}{Optional. Default ",".}
}
\description{
Convert string to character vector
}
\examples{
char.string2vector("a,b,c,d")
[1] "a" "b" "c" "d"
}
