% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{dedup_choice}
\alias{dedup_choice}
\title{Deduplication chooser
Majority rules, ties goes to lowest value after ordering, so most recent if a date or time.}
\usage{
dedup_choice(df)
}
\arguments{
\item{df}{Data.frame or data.table}
}
\description{
Deduplication chooser
Majority rules, ties goes to lowest value after ordering, so most recent if a date or time.
}
