% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime.R
\name{dttm.parse_dob_to_cols}
\alias{dttm.parse_dob_to_cols}
\title{Parse DOB to Separate Year, Month, Day Columns
Create separate year, month, day columns from ymd or mdy formatted date.}
\usage{
dttm.parse_dob_to_cols(df, date_field, date_format = "ymd")
}
\arguments{
\item{df}{data.frame or data.table with a date field.}

\item{date_field}{Name of the date field/column.}

\item{date_format}{Optional. Date field format. Either "ymd" (Year-Month-Day)
or "mdy" (Month-Day-Year). Defaults to "ymd".}
}
\description{
Parse DOB to Separate Year, Month, Day Columns
Create separate year, month, day columns from ymd or mdy formatted date.
}
\examples{
df_with_dob_column <- data.frame(dob = c("11/13/1968"))
df_with_parsed_dob <- dttm.parse_dob_to_cols(df_with_dob_column, date_field = "dob", date_format = "mdy")
df_with_parsed_dob
         dob birth_yr birth_mo birth_dy
1 11/13/1968     1968       11       13
}
