% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{file_unzip2temp}
\alias{file_unzip2temp}
\title{Unzip File to Temp Directory and Return Directory Path}
\usage{
file_unzip2temp(zipfilepath)
}
\arguments{
\item{zipfilepath}{Path to the zip file.}
}
\description{
Unzip File to Temp Directory and Return Directory Path
}
\examples{
file_unzip2temp("data/CENSUS/cb/cb_2016_01_bg_500k.zip")
[1] "/tmp/Rtmp8lE8Pj/"
}
