% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string.R
\name{get_after_last_separator}
\alias{get_after_last_separator}
\title{Return everything between after last occurence of separator and end of the string.}
\usage{
get_after_last_separator(string = "a.b?c.net", separator = ".",
  include_sep_in_output = FALSE)
}
\arguments{
\item{string}{A character string.}

\item{separator}{Default ".".}

\item{include_sep_in_output}{Optional. Include the separator in the output. Default FALSE.}
}
\description{
Return everything between after last occurence of separator and end of the string.
}
\examples{
str <- "I am smart enough, good looking enough, and gosh darn it; people like me."
get_after_last_separator(str, separator = ",")
[1] " and gosh darn it; people like me."
get_after_last_separator(str, separator = ",", include_sep_in_output = T)
[1] ", and gosh darn it; people like me."
}
