% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string.R
\name{get_before_first_separator}
\alias{get_before_first_separator}
\title{Return everything before the first occurence of provided separator.}
\usage{
get_before_first_separator(string = "", separator = ",",
  include_sep_in_output = FALSE)
}
\arguments{
\item{string}{A character string.}

\item{separator}{Optional. Default ",".}

\item{include_sep_in_output}{Optional. Include the separator in the output. Default FALSE.}
}
\description{
Return everything before the first occurence of provided separator.
}
\examples{
get_before_separator("I am smart enough, good looking enough, and gosh darn it, people like me.")
[1] "I am smart enough"
}
