% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string.R
\name{get_before_separator}
\alias{get_before_separator}
\title{Return everything before the first occurence of provided separator.}
\usage{
get_before_separator(string = "", separator = ",")
}
\arguments{
\item{string}{A character string.}

\item{separator}{Optional. Default ",".}
}
\description{
Return everything before the first occurence of provided separator.
}
\examples{
get_before_separator("I am smart enough, good looking enough, and gosh darn it, people like me.")
[1] "I am smart enough"
}
