% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string.R
\name{get_between_separators}
\alias{get_between_separators}
\title{Return everything between first occurence of first separator and first occurence
of second separator after the first separator.}
\usage{
get_between_separators(string = "", first_separator = ",",
  second_separator = ",")
}
\arguments{
\item{string}{A character string.}

\item{first_separator}{Optional. Default ",".}

\item{second_separator}{Optional. Default ",".}
}
\description{
Return everything between first occurence of first separator and first occurence
of second separator after the first separator.
}
\examples{
str <- "I am smart enough, good looking enough, and gosh darn it; people like me."
get_between_separators(str)
[1] "good looking enough"
get_between_separators(str, second_separator = ";")
[1] "good looking enough, and gosh darn it"
}
