% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{is_blank}
\alias{is_blank}
\title{Checks if provided variable is null, na, or of length 0}
\usage{
is_blank(x)
}
\arguments{
\item{x}{Variable provided.}
}
\description{
Checks if provided variable is null, na, or of length 0
}
\examples{
is_blank("")
[1] TRUE
}
