% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location.R
\name{loc.lat_lon2geo_areas}
\alias{loc.lat_lon2geo_areas}
\title{Get census block, county, state, and market area information based on latitude/longitude input.}
\usage{
loc.lat_lon2geo_areas(place_id = "VTRC", lat = 38.880807,
  lon = -77.11577)
}
\arguments{
\item{place_id}{some unique identifier for the lat lon}

\item{lat}{the latitude}

\item{lon}{the longitude}
}
\value{
data.frame
}
\description{
Get census block, county, state, and market area information based on latitude/longitude input.
}
\examples{
loc.lat_lon2geo_areas("VTRC", lat=38.880807, lon=-77.11577)
}
