% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location.R
\name{loc.lats_lons2geo_areas}
\alias{loc.lats_lons2geo_areas}
\title{Get census block, county, state, and market area information based on multiple latitude/longitude inputs.}
\usage{
loc.lats_lons2geo_areas(place_idCol = c("VTRC", "VT-NVC"),
  latCol = c(38.880807, 38.8968325), lonCol = c(-77.11577,
  -77.1894815))
}
\arguments{
\item{place_idCol}{vector of unique identifiers}

\item{latCol}{vector of latitudes}

\item{lonCol}{vector of longitudes}
}
\value{
data.frame
}
\description{
Get census block, county, state, and market area information based on multiple latitude/longitude inputs.
}
\examples{
loc.lats_lons2geo_areas(place_idCol = c("VTRC", "VT-NVC"),
                  latCol = c(38.880807, 38.8968325),
                  lonCol = c(-77.11577, -77.1894815))
}
