% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name.R
\name{make_names}
\alias{make_names}
\title{Make R compliant names that use an underscore rather than a dot.}
\usage{
make_names(char)
}
\description{
Make R compliant names that use an underscore rather than a dot.
}
\examples{
make_names(c("name with space", "name,with,commas"))
[1] "name_with_space"  "name_with_commas"
}
