% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name.R
\name{move_suffix}
\alias{move_suffix}
\title{Move provided suffixes to separate column (and remove suffix from name column).}
\usage{
move_suffix(df, name_field, suffix_list)
}
\arguments{
\item{df}{A data.frame or data.table with a suffix containing name column.}

\item{name_field}{Name of the name column/field.}

\item{suffix_list}{A list of suffixes to move.}
}
\description{
Move provided suffixes to separate column (and remove suffix from name column).
}
\examples{
names <- data.frame(name = c("Bob, Jr", "Dallas III", "Willy"))
names <- move_suffix(names, "name", list("Jr", "III"))
names
   name suffix
1  Bob,     Jr
2  Dallas    III
3  Willy   <NA>
}
