% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{new_row_per_string_item}
\alias{new_row_per_string_item}
\title{Create new data.table rows for each combination of first and last name}
\usage{
new_row_per_string_item(single_row_dt, fname_col, lname_col,
  delims = "[-, ]")
}
\arguments{
\item{single_row_dt}{Data.table with one row/record.}

\item{fname_col}{First name column name.}

\item{lname_col}{Last name column name.}

\item{delims}{Delimters to split string by.}
}
\description{
Create new data.table rows for each combination of first and last name
}
