% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime.R
\name{parse_dob_to_cols}
\alias{parse_dob_to_cols}
\title{Create separate year, month, day columns from ymd or mdy formatted date.}
\usage{
parse_dob_to_cols(df, date_field, date_format = "ymd")
}
\arguments{
\item{df}{data.frame or data.table with a date field.}

\item{date_field}{Name of the date field/column.}

\item{date_format}{Optional. Date field format. Either "ymd" (Year-Month-Day)
or "mdy" (Month-Day-Year). Defaults to "ymd".}
}
\description{
Create separate year, month, day columns from ymd or mdy formatted date.
}
\examples{
fix_century("2049-11-30")
[1] "1949-11-30"
}
