% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{set.rm_all_NA_rows}
\alias{set.rm_all_NA_rows}
\title{Remove rows that are completely NA}
\usage{
set.rm_all_NA_rows(df)
}
\arguments{
\item{df}{data.frame or data.table.}
}
\description{
Remove rows that are completely NA
}
\examples{
na_df <- data.frame(col1 = c("Hola", NA), col2 = c("Chicka", NA))
na_df <- set.rm_all_NA_rows(na_df)
na_df
   col1   col2
1: Hola Chicka
2:   NA     NA
}
