% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{set.split_hyphenated}
\alias{set.split_hyphenated}
\title{Split hyphenated strings in column into two columns}
\usage{
set.split_hyphenated(df, name_field, return_type = "dt")
}
\arguments{
\item{df}{data.frame or data.table.}

\item{name_field}{name of column containing hyphenated values.}

\item{return_type.}{Optional. Return data.table "dt" or data.frame "df". Default "dt".}
}
\description{
Split hyphenated strings in column into two columns
}
\examples{
df <- data.frame(first_name = c("Devante"), last_name = c("Smith-Pelly"))
dt <- set.split_hyphenated(df, "last_name")
dt
   first_name last_name last_name_left last_name_right
1:    Devante     Smith          Smith           Pelly
}
