% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recursion.R
\name{trampoline}
\alias{trampoline}
\title{Trampoline}
\usage{
trampoline(f, ...)
}
\arguments{
\item{f}{A function.}

\item{...}{Any number of named or unnamed arguments.}
}
\description{
A trampoline is a function or set of functions that together give us the tools we
need to write code in a recursive style, in a way that doesn't overflow the stack.
Here's an awesome trampoline by Jim Hester:
}
\examples{
countdown <- trampoline(function(n) {
  if (n > 0) recur(n-1) else "done"
})
countdown(10000)
[1] "done"
}
