% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/try.R
\name{try_try_try}
\alias{try_try_try}
\title{Try running a function and get a non-error response n times}
\usage{
try_try_try(x, n = 3L)
}
\arguments{
\item{x}{A function.}

\item{n}{Number of times to try.}
}
\description{
Try running a function and get a non-error response n times
}
