% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{var.is_blank}
\alias{var.is_blank}
\title{Checks if provided variable is null, na, or of length 0}
\usage{
var.is_blank(x)
}
\arguments{
\item{x}{Variable provided.}
}
\description{
Checks if provided variable is null, na, or of length 0
}
\examples{
var.is_blank("")
[1] TRUE
}
