% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string.R
\name{remove_non_alphanum}
\alias{remove_non_alphanum}
\title{Remove non-alphanumeric characters from a string.}
\usage{
remove_non_alphanum(string)
}
\arguments{
\item{string}{A character string.}
}
\description{
Remove non-alphanumeric characters from a string.
}
\examples{
remove_non_alphanum("222-44-6666")
[1] "222446666"
}
