% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string.R
\name{string2vector}
\alias{string2vector}
\title{Convert string to character vector}
\usage{
string2vector(string, delimeter = ",")
}
\arguments{
\item{string}{A character string.}

\item{delimeter}{Optional. Default ",".}
}
\description{
Convert string to character vector
}
\examples{
remove_non_alphanum("222-44-6666")
[1] "222446666"
}
